pyproject.toml
[project]
name = "mini-dj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = ["django>=5.1.3", "djangorestframework>=3.15.2"]

[dependency-groups]
dev = [
  "ipython>=8.29.0",
  "pyright>=1.1.388",
  "pytest-cov>=6.0.0",
  "pytest>=8.3.3",
  "ruff>=0.7.2",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN",
  "COM812",
  "D1",
  "D203",
  "D213",
  "EM101",
  "INP",
  "ISC001",
  "RUF012",
  "S101",
  "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**.py" = []

[tool.pytest.ini_options]
testpaths = ["tests"]
